+++ artiq/coredevice/core.py
@@ -106,8 +106,7 @@ class Core:
 
             module = Module(stitcher,
                 ref_period=self.ref_period,
-                attribute_writeback=attribute_writeback,
-                raise_assertion_errors=self.target_cls.raise_assertion_errors)
+                attribute_writeback=attribute_writeback)
             target = self.target_cls()
 
             library = target.compile_and_link([module])